+++ CHANGELOG.md
@@ -1,23 +1,31 @@
 # CHANGELOG
+===================
+
 
 ## v5.0-r1 (2025-05-02)
+-----------
+* Moodle 5.0 compatible version
 * [FEATURE] Add further behat tests and update existing ones
 * [UPDATE] Update of readme file
-* Moodle 5.0 compatible version
 
 ## 4.5.0
+-----------
 * Moodle 4.5 compatible version
 
 ## 4.4.0
+-----------
 * Moodle 4.4 compatible version
 
 ## 4.3.0
+-----------
 * Moodle 4.3 compatible version
 
 ## 4.2.0
+-----------
 * Moodle 4.2 compatible version
 
 ## 4.1.0
+-----------
 * Moodle 4.1 compatible version
 * [FIX] Sorting of columns
 * [FIX] Message column will not be rendered when messaging is disabled
@@ -25,15 +33,18 @@
 * [UPDATE] Removed unused code
 * [FEATURE] Added some behat tests
 
-
 ## 4.0.1
+-----------
 * fix a bug where users changing their option to "no" will never be displayed for case-sensitive databases
 
 ## 4.0.0
+-----------
 * Moodle 4.0 compatible version
 
 ## 3.11.0
+-----------
 * Moodle 3.11 compatible version
 
 ## 3.10.0
+-----------
 * Moodle 3.10 compatible version